chore(node/tests): deprecate devnet-sdk
