+++ .github/workflows/e2e_tests.yaml
@@ -71,9 +71,6 @@ jobs:
       - name: test node with ${{ matrix.devnet-config }} devnet
         run: just test-e2e "${{ matrix.devnet-config }}" node
 
-      - name: test node-devstack with ${{ matrix.devnet-config }} devnet
-        run: just test-e2e "${{ matrix.devnet-config }}" node-devstack
-
       - # Temp fix
         # https://github.com/docker/build-push-action/issues/252
         # https://github.com/moby/buildkit/issues/1896
