+++ tests/README.md
@@ -1,6 +1,6 @@
 # E2e testing for the kona-node
 
-This repository contains the e2e testing resources for the kona-node. The e2e testing is done using the `devnet-sdk` from the [Optimism monorepo](https://github.com/ethereum-optimism/optimism). For now, only deployments with kurtosis are supported. The `devnet-sdk` is a framework that provides bindings to optimism devnets to help writing e2e tests.
+This repository contains the e2e testing resources for the kona-node. The e2e testing is done using the `devstack` from the [Optimism monorepo](https://github.com/ethereum-optimism/optimism). For now, only deployments with kurtosis are supported. The `devstack` is a framework that provides bindings to optimism devnets to help writing e2e tests.
 
 ## Installation
