+++ tests/node/init_test.go
@@ -1,4 +1,4 @@
-package nodedevstack
+package node
 
 import (
 	"testing"
