+++ tests/node/mixed_preset.go
@@ -1,4 +1,4 @@
-package nodedevstack
+package node
 
 import (
 	"fmt"
